chore: move plugin manager module to core
