+++ packages/core/src/provider/index.ts
@@ -1,7 +1,7 @@
 export { rspackProvider } from './provider';
 export type { Rspack, RspackConfig } from '@rsbuild/shared';
 export { createContext, createPublicContext } from './core/createContext';
-export { initPlugins } from '@rsbuild/shared';
+export { initPlugins, createPluginManager } from '../pluginManager';
 export { initHooks, type Hooks } from './core/initHooks';
 export { initRsbuildConfig } from './core/initConfigs';
 export { getPluginAPI } from './core/initPlugins';
