--- packages/shared/tests/createPluginStore.test.ts
@@ -1,26 +1,0 @@
-import { createPluginStore } from '../src/pluginStore';
-
-describe('createPluginStore', () => {
-  it('addPlugins and removePlugins works', () => {
-    const pluginStore = createPluginStore();
-    expect(pluginStore.plugins).toEqual([]);
-    pluginStore.addPlugins([
-      {
-        name: 'foo',
-        setup() {
-          /* do nothing */
-        },
-      },
-      {
-        name: 'bar',
-        setup() {
-          /* do nothing */
-        },
-      },
-    ]);
-    expect(pluginStore.plugins).toHaveLength(2);
-    pluginStore.removePlugins(['foo']);
-    expect(pluginStore.plugins).toHaveLength(1);
-    expect(pluginStore.isPluginExists('foo')).toBe(false);
-  });
-});
